added migrations and removed functions on the components

+++ database/migrations/2025_08_05_130813_create_vehicles_table.php
@@ -11,26 +11,29 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('vehicles', function (Blueprint $table) {
-            $table->id();
-            $table->string('plate_number', 20)->unique();
-            $table->string('vehicle_make', 50);
-            $table->string('vehicle_model', 50);
-            $table->string('vehicle_color', 30);
-            $table->enum('vehicle_type', ['car', 'suv', 'truck', 'van'])->default('car');
-            $table->string('rfid_tag', 50)->unique();
-            $table->unsignedBigInteger('owner_id');
-            $table->datetime('expires_at')->nullable();
-            $table->boolean('is_active')->default(true);
-            $table->timestamps();
+        if (!Schema::hasTable('vehicles')){
+                Schema::create('vehicles', function (Blueprint $table) {
+                $table->id();
+                $table->string('plate_number', 20)->unique();
+                $table->string('vehicle_make', 50);
+                $table->string('vehicle_model', 50);
+                $table->string('vehicle_color', 30);
+                $table->enum('vehicle_type', ['car', 'suv', 'truck', 'van'])->default('car');
+                $table->string('rfid_tag', 50)->unique();
+                $table->unsignedBigInteger('owner_id');
+                $table->datetime('expires_at')->nullable();
+                $table->boolean('is_active')->default(true);
+                $table->timestamps();
 
-            $table->index('plate_number');
-            $table->index('rfid_tag');
-            $table->index('owner_id');
-            $table->index('expires_at');
-            $table->index('is_active');
-            $table->foreign('owner_id')->references('id')->on('sys_users')->onDelete('cascade');
-        });
+                $table->index('plate_number');
+                $table->index('rfid_tag');
+                $table->index('owner_id');
+                $table->index('expires_at');
+                $table->index('is_active');
+                $table->foreign('owner_id')->references('id')->on('sys_users')->onDelete('cascade');
+            });
+        }
+        
     }
 
     /**
